console: Move views to separate folder
