+++ pkg/webui/console/views/app/index.js
