+++ pkg/webui/console/views/application-access-list/index.js
