+++ pkg/webui/console/views/application-add/application-add.styl
