+++ pkg/webui/console/views/application-add/index.js
