+++ pkg/webui/console/views/application-general-settings/application-general-settings.styl
