+++ pkg/webui/console/views/application-general-settings/index.js
