+++ pkg/webui/console/views/application-overview/application-overview.styl
