+++ pkg/webui/console/views/application-overview/index.js
