+++ pkg/webui/console/views/application/index.js
