+++ pkg/webui/console/views/applications-list/index.js
