+++ pkg/webui/console/views/applications/index.js
