+++ pkg/webui/console/views/gateways-list/index.js
