+++ pkg/webui/console/views/gateways/index.js
