+++ pkg/webui/console/views/landing/index.js
