+++ pkg/webui/console/views/overview/index.js
